Move NodeJS launcher to a separated folder
