+++ .gitignore
@@ -1,6 +1,7 @@
 .DS_Store
-/node_modules
+node_modules
 *-lock.*
 *.lock
 *.log
 dist
+src/package.json
