+++ node/setup.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+npm install ws
+echo '{"type": "module"}' > ../src/package.json
